replace database with isomorphicdb wherever appropriate
