+++ .github/workflows/release.yml
@@ -24,5 +24,5 @@ jobs:
           username: ${{ github.actor }}
           password: ${{ secrets.GITHUB_TOKEN }}
           registry: docker.pkg.github.com
-          repository: alex-dukhno/database/database
+          repository: alex-dukhno/isomorphicdb/isomorphicdb
           tag_with_ref: true
