+++ docs/user-guide/README.md
@@ -1,3 +1,5 @@
 # User Guide
 
+## :warning: Currently Query Engine is under massive redevelopment. Following instructions will not work :warning:
+
 ### [Getting Started](getting_started.md)
