+++ docs/user-guide/getting_started.md
@@ -1,6 +1,6 @@
 # Getting Started
 
-Currently, `database` is not distributed in a binary format. Experimentally, we
+Currently, `isomorphicdb` is not distributed in a binary format. Experimentally, we
 support docker image. So you have to have Docker installed on your machine.
 Before pooling it from `GitHub` registry you have to perform the following steps:
 
@@ -19,25 +19,25 @@ cat /path/to/file/with/token | docker login https://docker.pkg.github.com -u <yo
 After that you can pool image with the following command:
 
 ```shell script
-docker pull docker.pkg.github.com/alex-dukhno/database/database:latest
+docker pull docker.pkg.github.com/alex-dukhno/isomorphicdb/isomorphicdb:latest
 ```
 
 To start up application you need to invoke the following command:
 
 ```shell script
-docker run -it -d -p 5432:5432 docker.pkg.github.com/alex-dukhno/database/database
+docker run -it -d -p 5432:5432 docker.pkg.github.com/alex-dukhno/isomorphicdb/isomorphicdb
 ```
 
 If you crashed the database docker instance we highly appreciated if you rerun
 scenario with the following command and post a bug with backtrace:
 
 ```shell script
-docker run -it -d -e RUST_BACKTRACE=1 -p 5432:5432 docker.pkg.github.com/alex-dukhno/database/database
+docker run -it -d -e RUST_BACKTRACE=1 -p 5432:5432 docker.pkg.github.com/alex-dukhno/isomorphicdb/isomorphicdb
 ```
 Thanks!
 
 To connect to database you have to have `psql` on your machine, it can be installed
-with `PostgreSQL` from the https://www.postgresql.org[official website] or with
+with `PostgreSQL` from the [official website](https://www.postgresql.org) or with
 package manager like `homebrew` or `apt-get`.
 
 Then you can start client with the command:
