+++ tests/compatibility/src/test/groovy/io/isomorphicdb/extended/BasicQueryOperationsSpec.groovy
@@ -1,17 +1,26 @@
-package io.database.simple
+package io.isomorphicdb.extended
 
 import groovy.sql.GroovyRowResult
-import io.database.ThreeSmallIntColumnTable
+import io.isomorphicdb.ThreeSmallIntColumnTable
 
 class BasicQueryOperationsSpec extends ThreeSmallIntColumnTable {
-  def 'insert select{all}'() {
-    given:
-      String insertQuery = 'insert into SCHEMA_NAME.TABLE_NAME values (1, 2, 3), (4, 5, 6), (7, 8, 9)'
+  private static String INSERT_QUERY = 'insert into SCHEMA_NAME.TABLE_NAME values (?, ?, ?)'
 
+  private int pgInserts = 0
+  private int dbInserts = 0
+
+  def setup() {
+    pgInserts += pg.executeUpdate INSERT_QUERY, [1, 2, 3]
+    pgInserts += pg.executeUpdate INSERT_QUERY, [4, 5, 6]
+    pgInserts += pg.executeUpdate INSERT_QUERY, [7, 8, 9]
+
+    dbInserts += db.executeUpdate INSERT_QUERY, [1, 2, 3]
+    dbInserts += db.executeUpdate INSERT_QUERY, [4, 5, 6]
+    dbInserts += db.executeUpdate INSERT_QUERY, [7, 8, 9]
+  }
+
+  def 'insert select{all}'() {
     when:
-      int pgInserts = pg.executeUpdate insertQuery
-      int dbInserts = db.executeUpdate insertQuery
-    and:
       List<GroovyRowResult> pgSelect = pg.rows SELECT_ALL_QUERY
       List<GroovyRowResult> dbSelect = db.rows SELECT_ALL_QUERY
 
@@ -25,13 +34,9 @@ class BasicQueryOperationsSpec extends ThreeSmallIntColumnTable {
 
   def 'insert select{listed column}'() {
     given:
-      String insertQuery = 'insert into SCHEMA_NAME.TABLE_NAME values (1, 2, 3), (4, 5, 6), (7, 8, 9)'
       String selectQuery = 'select col1, col2, col3 from SCHEMA_NAME.TABLE_NAME'
 
     when:
-      int pgInserts = pg.executeUpdate insertQuery
-      int dbInserts = db.executeUpdate insertQuery
-    and:
       List<GroovyRowResult> pgSelect = pg.rows selectQuery
       List<GroovyRowResult> dbSelect = db.rows selectQuery
 
@@ -45,18 +50,14 @@ class BasicQueryOperationsSpec extends ThreeSmallIntColumnTable {
 
   def 'insert update{all} select{all}'() {
     given:
-      String insertQuery = 'insert into SCHEMA_NAME.TABLE_NAME values (1, 2, 3), (4, 5, 6), (7, 8, 9)'
-      String updateQuery = 'update SCHEMA_NAME.TABLE_NAME set col1 = 10, col2 = 11, col3 = 12'
-    and:
-      pg.executeUpdate insertQuery
-      db.executeUpdate insertQuery
+      String updateQuery = 'update SCHEMA_NAME.TABLE_NAME set col1 = ?, col2 = ?, col3 = ?'
 
     when:
-      int pgUpdates = pg.executeUpdate updateQuery
-      int dbUpdates = db.executeUpdate updateQuery
+      int pgUpdates = pg.executeUpdate updateQuery, [10, 11, 12]
+      int dbUpdates = db.executeUpdate updateQuery, [10, 11, 12]
     and:
       List<GroovyRowResult> pgSelect = pg.rows SELECT_ALL_QUERY
-      List<GroovyRowResult> dbSelect =db.rows SELECT_ALL_QUERY
+      List<GroovyRowResult> dbSelect = db.rows SELECT_ALL_QUERY
 
     then:
       println "UPDATED: ${pgUpdates.inspect()}"
@@ -68,11 +69,7 @@ class BasicQueryOperationsSpec extends ThreeSmallIntColumnTable {
 
   def 'insert delete{all} select{all}'() {
     given:
-      String insertQuery = 'insert into SCHEMA_NAME.TABLE_NAME values (1, 2, 3), (4, 5, 6), (7, 8, 9)'
       String deleteQuery = 'delete from SCHEMA_NAME.TABLE_NAME'
-    and:
-      pg.executeUpdate insertQuery
-      db.executeUpdate insertQuery
 
     when:
       int pgDeletes = pg.executeUpdate deleteQuery
