+++ tests/compatibility/src/test/groovy/io/isomorphicdb/simple/SelectiveAttributesQueryOperationsSpec.groovy
@@ -1,7 +1,7 @@
-package io.database.simple
+package io.isomorphicdb.simple
 
 import groovy.sql.GroovyRowResult
-import io.database.ThreeSmallIntColumnTable
+import io.isomorphicdb.ThreeSmallIntColumnTable
 
 class SelectiveAttributesQueryOperationsSpec extends ThreeSmallIntColumnTable {
   private static final String INSERT_QUERY = '''
